python/requirements.txt
paho-mqtt==1.6.1
influxdb-client==1.43.0
flask==3.0.0
flask-cors==4.0.0
